refactor: rename identitySlice name to registerIdentitySlice
