+++ src/components/header/NavigationDefault.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
-import { setIdentityState } from "../../../redux/common/identitySlice";
+import { setIdentityState } from "../../../redux/common/registerIdentitySlice";
 
 function NavigationDefault() {
   const navigate = useNavigate();
