+++ src/pages/login/loginPage.tsx
@@ -4,19 +4,25 @@ import LandLordLogin from "../../components/login/landLordLogin";
 import { RootState } from "../../../redux/store";
 
 function LoginPage() {
-  const identityState = useSelector((store: RootState) => store.identityState.identity);
-  console.log(identityState);
+  const registerIdentityState = useSelector(
+    (store: RootState) => store.registerIdentityState.identity
+  );
+  console.log(registerIdentityState);
   return (
     <div
       className={
-        identityState === "tenant"
+        registerIdentityState === "tenant"
           ? "wrap h-screen bg-Neutral-99 bg-tenantLoginImg bg-contain bg-no-repeat"
           : "wrap h-screen bg-Neutral-99 bg-landLordLoginImg bg-contain bg-no-repeat"
       }
     >
       <div className="container layout-grid pt-[170px]">
         <div className="col-span-5 col-start-7">
-          {identityState === "tenant" ? <TenantLogin /> : <LandLordLogin />}
+          {registerIdentityState === "tenant" ? (
+            <TenantLogin />
+          ) : (
+            <LandLordLogin />
+          )}
         </div>
       </div>
     </div>
